requirements.txt
fastapi
uvicorn[standard]

langchain
langchain-google-genai
langchain-community

pymongo
pymongo[srv]
pydantic>=2.0.0
python-dotenv
requests

# For reading PDFs & chunking
pypdf
chromadb

# Optional but recommended for embeddings
sentence-transformers


transformers
torch
pymongo